Convert integration tests into regular tests
The plugins (existing and planned ones) in this repository are not
likely to have any platform-specific bits. This means that the sources
will be pure-Python implementations, and the elements will _not_ be
Build Elements, i.e. won't need to run sandbox commands. For example,
the Docker source plugin doesn't have any host dependencies, and the OCI
element plugin will simply be collecting the required files in a specifc
directory structure.