setup.py, .gitlab-ci.yml: Remove coverage steps
The coverage reporting is already broken for this repository for a
couple for reasons. We are missing a `.coveragerc` file and more
importantly, coverage.py is not capturing any data even after addding a
config file. I suspect this is because of the way we run our tests via
invoking the CLI, and not by importing our module directly.

In future, we may find other ways to work around this issue, and we can
look into re-enabling the coverage checks. But, it makes sense to remove
these until then as they are not adding any value at present.