setup.py,.gitlab-ci.yml: Run tests using tox
Following the direction the main BuildStream repo is headed, move to
running tests using `tox`. After doing this, we get the following
benefits:

- No need to hackily install our runtime dependencies, this is handled
  by tox now.
- No need for a separate "sdist" stage in the CI.

But, this means that `setup.py test` command won't work anymore as
we don't list the testing dependencies as `tests_requires`, but that
rather as "extras" under the name "test". This will now allow folks to
install the test dependencies locally using something like
`pip3 install .[test]`.