.gitlab-ci.yml: Always build docs, only publish from master
Currently, we only build docs when running on the master branch, which
makes it harder to catch any issues with the docs before the build. With
this change, we will always run the docs build, but only publish when
running from the master branch.

Also, move the build dependencies for documentation as an "extra" in
`setup.py` to ease the installation process on CI and developer laptops.