tests: Remove testutils/plugin_import.py
This file was originally added in
https://gitlab.com/BuildStream/bst-external/merge_requests/46 to
avoid loading the module using pip. But since we run our tests using
`tox` now, we do not have that limitation anymore. So, we can get rid of
this hack. We do not need this hack anymore as we run our tests using
`tox` now.

.gitlab-ci.yml: Upgrade pip before running tests. Old versions of pip
(especially the default on python:3.5-stretch image) cause issues with
our tests.