.gitlab-ci.yml: Remove tox-venv
tox-env avoids using virtualenv, and instead uses the venv module from
the standard library. venv, by design, uses a bundled version of pip.
This version can sometimes be quite old. For example,
https://gitlab.com/BuildStream/bst-plugins-container/-/jobs/229329441
was failing because the version of pip inside the virtual environment
was 9.0.1, which was rather old now.

Fixes #7.