tests/docker_element.py: Make file duplication assertion stronger
Previously the logic for calculating file duplications was incorrect, as
it was looking for files that were present in all layers rather doing a
pair-wise comparison. Additionally, test assertions have now been
updated to include number of files that are shared between layers.
This is particularly important for `test_nested_overwrite` as now we
explicitly assert that files are present in both layers.