docker_image: Use os.walk to add files to Docker layers
Previously, py36. py35 tests were failing due to the fact that:
- `tar.add(recurse=True)` was sorting files in py37 but not in other
versions. This was causing Docker layers' SHAs to be named differently
between the versions.
- `test_nested_deps_docker_image` was extracting all layers. Thus
assertions on file contents were placed on the last layer extracted by
`os.listdir()`, which was an arbritary layer due to non-determinisic layer
SHAs, rather than on the final layer of the Docker image.