Use black to format source code
Start using black (https://github.com/psf/black) as code formatter. This
is mainly so that we don't have to waste time worrying about the style
of the code. Letting `black` take care of the boring stuff will allow us
to focus on the writing the code.

`black` also offers integrations for various editors and IDEs. Check
its GitHub page for more details.