tests/docker_element.py: Fix failing test_share_layers_docker_image
This test creates two images and checks that the two images share the
same base layer. Testsi were failing because BuildStream throws away modification
times when caching, hence the modification time for files when adding to
the layer tarball for Docker images is the time when they are staged.
This means that even if two images use the same file-system for a layer,
the layer's SHA will not be identical as each layer's files will have
the timestamp correpsonding to it's staging time, hence the two images
will not be able to share the layer.

Proposed Patch
- Use user-specified timestamp for mtime for files in layer tarballs.
- Ensure images built for this test use `deterministic` timestamp option