sources/docker.py: Handle fetch network error gracefully
The Docker source first pulls the manifest of the image and then pulls
the layers associated with the manifest. Some network errors might occur
after the manifest is pulled but before the associated layers are
pulled. In such casses, `sources/docker.py:get_consistency` will attempt
to verify a blob that does not exist. This will throw a `OS Error`.

Fixes issue #16