docker_image.py: Fix/silence pylint warnings
For each pylint warning, either fix it or disable it.

Other than the obvious changes, I have removed the `_pack_image()`
method. It was only used in one place, and it seems unlikely it will be
needed in other places. The body of the method has been moved to the
`assemble()` method which was the only caller.