sources/docker: Fix/silence pylint warnings
`requests.codes` provides access to its members as a regular dictionary,
as well as allows them to be referenced by name. pylint seems to dislike
the latter. This is really a pylint problem, but changing the style to
dictionary-like access seems less bad than adding a comment to disable
it.