docker_image: Ensure same image can have multiple tags
Converting the list of full names to a mapping (name -> tag) is wrong as
it's perfectly valid to supply two tags for the same name, like
`python:latest`, `python:3.7` and so on.

Change the code to store the names as tuples instead of dictionaries.
Also, add a regression test for this to ensure we don't break this in
future.

Fixes #18.